+++ src/plugins/intersection.script.js
@@ -1,0 +1,4 @@
+import {svgPlugin} from './intersection.js';
+if (gSVG) {
+  gSVG.install(svgPlugin);
+}